broke out all the schemas into models.js and created controllers for all the requestfunctionality under folder controllers
