+++ frontend/src/utils/utils.js
@@ -1,5 +1,5 @@
-const BASE_URL = "https://project-auth-kyvbhopdlq-lz.a.run.app";
-// const BASE_URL = "http://localhost:8090";
+//const BASE_URL = "https://project-auth-kyvbhopdlq-lz.a.run.app";
+const BASE_URL = "http://localhost:8090";
 const HAPPY_URL = 'https://project-happy-thoughts-api-dxrcv2y6yq-lz.a.run.app'
